Pixel perfect settings
Implemented various pixel perfect settings.
Most important: Canvas must use a separate camera to render the UI.
Therefore, every scene now has 2 cameras.

Since this is a lot of changes that mess with the formatting and everything, I've put it in a different branch. Programming team, if you would, please check that everything looks OK and if so merge it into the main branch.